# application/backend/certiblock/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/universities": {
            "get": {
                "description": "Get all universities",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Get all universities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.University"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/universities/certificate-file/approve": {
            "post": {
                "description": "Duyệt một chứng chỉ và đưa lên blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificates"
                ],
                "summary": "Duyệt certificate lên blockchain",
                "parameters": [
                    {
                        "description": "Thông tin duyệt certificate",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.BlockchainCertificateOutput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/universities/certificate-file/request": {
            "post": {
                "description": "Student issue a certificate file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Student issue a certificate file",
                "parameters": [
                    {
                        "description": "Certificate file data",
                        "name": "certificateFile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.CertificateFileUpload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/universities/certificates/not_on_chain": {
            "get": {
                "description": "Trả về các chứng chỉ chưa được đưa lên blockchain bởi các trường đại học",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificates"
                ],
                "summary": "Lấy danh sách certificate chưa được đưa lên blockchain",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.CertificateFileOutput"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/universities/certificates/on_chain": {
            "get": {
                "description": "Trả về các chứng chỉ đã được đưa lên blockchain bởi các trường đại học",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificates"
                ],
                "summary": "Lấy danh sách certificate đã được đưa lên blockchain",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.CertificateFileOutput"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/universities/register": {
            "post": {
                "description": "Register a university",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Register a university",
                "parameters": [
                    {
                        "description": "University registration data",
                        "name": "university",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UniversityInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.UniversityOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/universities/{id}": {
            "get": {
                "description": "Retrieve a university by its unique ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Get a university by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "University ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.University"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.BlockchainCertificateOutput": {
            "type": "object",
            "properties": {
                "certHash": {
                    "type": "string"
                },
                "certUUID": {
                    "type": "string"
                },
                "dateOfIssuing": {
                    "type": "string"
                },
                "studentPK": {
                    "type": "string"
                },
                "studentSignature": {
                    "type": "string"
                },
                "universityPK": {
                    "type": "string"
                },
                "universitySignature": {
                    "type": "string"
                }
            }
        },
        "data.CertificateFileOutput": {
            "type": "object",
            "properties": {
                "cert_uuid": {
                    "type": "string"
                },
                "date_of_issue": {
                    "type": "string"
                },
                "is_on_chain": {
                    "type": "boolean"
                },
                "plain_text_file_data": {
                    "type": "string"
                },
                "student_public_key": {
                    "type": "string"
                },
                "student_signature": {
                    "type": "string"
                },
                "university_name": {
                    "type": "string"
                },
                "university_public_key": {
                    "type": "string"
                },
                "university_signature": {
                    "type": "string"
                }
            }
        },
        "data.CertificateFileUpload": {
            "type": "object",
            "properties": {
                "certUUID": {
                    "type": "string"
                },
                "dateOfIssue": {
                    "type": "string"
                },
                "encodedFile": {
                    "type": "string"
                },
                "isOnChain": {
                    "type": "boolean"
                },
                "studentPublicKey": {
                    "type": "string"
                },
                "studentSignature": {
                    "type": "string"
                },
                "universityName": {
                    "type": "string"
                },
                "universitySignature": {
                    "type": "string"
                }
            }
        },
...
        "data.University": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "data.UniversityInput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "privateKey": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                }
            }
        },
        "data.UniversityOutput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "privateKey": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        }
    }
}